flask-server/requirements.txt
Flask
pika
Pillow
gunicorn
opencv-python-headless
python-dotenv
numpy
requests